requirements.txt
# requirements.txt
# 專案所需 Python 套件（建議使用 Python 3.9 ~ 3.12）

azure-identity>=1.18.0          # 認證相關（DefaultAzureCredential, ClientSecretCredential）
azure-mgmt-subscription>=3.0.0  # 取得 Azure 訂閱清單
azure-mgmt-costmanagement>=4.0.0  # 查詢成本管理 API（Defender 費用）
pandas>=2.0.0                   # 資料處理與 Excel 輸出
openpyxl>=3.0.0                 # pandas 寫入 xlsx 時的引擎